refactor : socket 분리를 위한 리펙토링
